+++ docs/changelog.rst
@@ -78,6 +78,7 @@ Changelog
 * :feature:`9972` rotki will now properly handle Quickswap transactions and LP balances in all supported chains.
 * :feature:`-` Users will now be able to apply custom accounting rules to a specific event or group of related events without affecting existing rules.
 * :feature:`10321` Users can now delete multiple events in the history events section.
+* :feature:`9133` rotki will now properly decode wrapping stETH into wstETH and vice versa.
 * :feature:`-` Giveth donation transactions on Polygon PoS will now be properly decoded.
 * :feature:`-`: Uniswap V2 transactions will now be decoded correctly on all supported chains rather than just on mainnet.
 * :feature:`-`: Users will now be able to search for a chain when adding a new evm chain asset.